PIPELINE/requirements.txt
faiss_cpu==1.9.0
NeighborBlend==0.3.0
numpy==1.23.0
onnx_graphsurgeon==0.5.8
opencv_contrib_python==4.6.0.66
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.82
psutil==5.9.8
pycuda==2024.1
pynvml==12.0.0
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.7.0
tensorrt==8.6.1
torch==2.7.0+cu118
torchvision==0.22.0+cu118
ultralytics==8.3.146
